backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
numpy==1.24.3
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
xgboost==2.0.3
shap==0.46.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
tqdm==4.66.1
python-multipart==0.0.7
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
reportlab==4.2.2
